Add basic lexer
Parsing all of the tokens manually, including keeping track of
whitespace and comments, was going to be very unwieldy. This commit adds
a basic lexer to handle the tokenizing.

In that process, I have found that the lexer has some context. This'll
be a pain. I will have to fix that later.